requirements.txt
transformers
pandas
accelerate
bitsandbytes
evaluate
datasets
python-dotenv

#metrics
peft
rouge-score
bert-score